NE64BADGE: Rename files for better conformance to naming conventions
